Revert "Redid folder structure"
