--- src/scenes/App/styles.scss
@@ -1,28 +1,0 @@
-// app.scss
-@import "scss/main.scss";
-
-#app,
-#body,
-#body > div {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-}
-
-#app {
-  min-height: 100vh;
-  background-color: whitesmoke;
-}
-
-#body:after {
-  content: ‘\00a0’;
-  display: block;
-  margin-top: var(— space);
-  height: 0;
-  visibility: hidden;
-}
-@media (min-width: 576px) {
-  #body {
-    margin: 0;
-  }
-}
