--- src/scenes/Home/components/PictureRow/styles.scss
@@ -1,6 +1,0 @@
-.picture {
-  background-size: cover;
-  height: 250px;
-  margin: 10px;
-  width: 250px;
-}
