--- src/scenes/Shop/components/shirtShop.js
@@ -1,24 +1,0 @@
-import React, {Component} from 'react';
-
-export class ShirtShop extends Component {
-  render() {
-    return (<div>
-      <div className="row">
-        <div className="col">
-          <p>Shirt</p>
-        </div>
-        <div className="col">
-          <p>Shirt</p>
-        </div>
-      </div>
-      <div className="row">
-        <div className="col">
-          <p>Shirt</p>
-        </div>
-        <div className="col">
-          <p>Shirt</p>
-        </div>
-      </div>
-    </div>)
-  }
-}
