+++ src/scenes/home.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import {PictureRow} from './components/PictureRow';
-import './styles.css';
+import {PictureRow} from 'components/picture-row';
+import 'scss/scenes/home.css';
 
 export class Home extends Component {
   render() {
