+++ src/scenes/lookbook.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import {ComingSoon} from '../../components/ComingSoon';
-import './styles.css';
+import {ComingSoon} from '../components/coming-soon';
+import 'scss/scenes/lookbook.css';
 
 export class Lookbook extends Component {
   render() {
