+++ src/scenes/shop.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import {ComingSoon} from '../../components/ComingSoon';
-import './styles.css';
+import { ComingSoon } from '../components/coming-soon';
+import 'scss/scenes/shop.css';
 
 export class Shop extends Component {
   render() {
