+++ src/scss/components/_all.scss
@@ -1,0 +1,3 @@
+@import 'footer.scss';
+@import 'navbar.scss';
+@import 'picture-Row.scss';
