+++ src/scss/components/_picture-Row.scss
@@ -1,0 +1,6 @@
+.picture {
+    background-size: cover;
+    height: 250px;
+    width: 250px;
+    margin: 10px;
+}
