+++ src/scss/main.scss
@@ -1,5 +1,9 @@
 // Modules and Variables
 @import 'scss/partials/base';
+
 // Partials
 @import 'scss/partials/typography';
 @import 'scss/partials/figures';
+
+// Components
+@import 'scss/components/all';
