+++ src/scss/partials/_typography.scss
@@ -1,16 +1,16 @@
 html {
-  font-family: $base-font-family;
-  @include font-size-map($body-font-size);
+    font-family: $base-font-family;
+    @include font-size-map($body-font-size);
 }
 
 h1 {
-  font-family: $header-font-family;
-  @include font-size-map($h1-font-size);
+    font-family: $header-font-family;
+    @include font-size-map($h1-font-size);
 }
 
 h2 {
-  font-family: $header-font-family;
-  @include font-size-map($h2-font-size);
+    font-family: $header-font-family;
+    @include font-size-map($h2-font-size);
 }
 
 .NavLink {
