+++ src/scss/scenes/app.scss
@@ -1,0 +1,29 @@
+// app.scss
+
+@import "scss/main.scss";
+#app,
+#body,
+#body>div {
+	flex: 1;
+	display: flex;
+	flex-direction: column;
+}
+
+#app {
+	min-height: 100vh;
+	background-color: whitesmoke;
+}
+
+#body:after {
+	content: ‘\00a0’;
+	display: block;
+	margin-top: var(— space);
+	height: 0;
+	visibility: hidden;
+}
+
+@media (min-width: 576px) {
+	#body {
+		margin: 0;
+	}
+}
